requirements.txt
mcpi
pillow
